database part migrated to sqldelight
creating invoices is completed

+++ src/jvmMain/kotlin/uz/qmgroup/budget_accounting/datasource/models/Person.kt
@@ -1,7 +1,7 @@
 package uz.qmgroup.budget_accounting.datasource.models
 
 data class Person(
-    val id: Int? = null,
+    val id: Long? = null,
     val name: String,
-    val balance: Float,
+    val balance: Double,
 )
